Typing `DescriptionBox`, cleaning up unused functionality.
